# POC: optimize small values encoding