This is not a pull request, more of a discuss request :)

We have a piece of software that accepts data packages of around 1k to 10k bytes and stores them for some time. In order to save some RAM we of course compress them. Current implementation uses Snappy for compression, but we thought Zstd with a dictionary would be a nice replacement - the incoming data packages mostly contain repeating words.
I tried replacing Snappy with Zstd Go implementation (EncodeAll() + a custom dict), but the overall performance was mediocre, even though compression level was very impressive indeed.

After a short investigation I found that most of the time is spent in copy()'ing dictionary table to the one used by the encoder. I'm not really proficient in compression algorithms, but from what I saw in the C sources, it's just a part of Zstd design, no way to get rid of it completely. In the same time there is no much sense in copying 32K table (for fast, it's 128K for default) because some of its elements were changed. Quite obviously this copy() call takes more time than the compression of a 1k package itself.

Hence the following patch: it introduces "virtual shards" in the table and marks the updated ones as dirty, so that they could be re-initialized with clean dictionary data later. 

**This is just a proof of concept**, I don't like the modified code myself, it's quite dumb, there has to be a cleaner way of doing it, but it seems that each "cleaner" version I tried affects performance in a very noticeable way, so let's just start with this one. 
I also modified the benchmark suite a bit in order to make it more convenient (and shorter): divided the benchmarks into several categories by data size.

Apparently, this improvement mainly for small data packages, bigger data takes more time to compress and the table shards will be mostly 'dirty', so there's no way to avoid copy()'ing all the table after all.

Let me know what you think about it.

See some benchmarks below.

Data size less than 1024 bytes:
```
EncodeAllDict0_1024/length-19-level-fastest-dict-1-8        9.33µs ± 1%     1.08µs ± 1%    -88.47%  (p=0.000 n=10+9)
EncodeAllDict0_1024/length-19-level-default-dict-1-8        46.0µs ± 2%      1.6µs ± 1%    -96.46%  (p=0.000 n=10+10)
EncodeAllDict0_1024/length-19-level-better-dict-1-8          173µs ± 6%      178µs ± 3%     +3.06%  (p=0.029 n=10+10)
EncodeAllDict0_1024/length-5-level-fastest-dict-1-8         9.34µs ± 1%     0.92µs ± 0%    -90.09%  (p=0.000 n=10+9)
EncodeAllDict0_1024/length-5-level-default-dict-1-8         45.5µs ± 2%      1.3µs ± 1%    -97.14%  (p=0.000 n=10+9)
EncodeAllDict0_1024/length-5-level-better-dict-1-8           167µs ± 2%      170µs ± 1%     +1.34%  (p=0.011 n=9+9)
EncodeAllDict0_1024/length-659-level-fastest-dict-1-8       16.8µs ± 2%     11.4µs ± 1%    -31.76%  (p=0.000 n=10+10)
EncodeAllDict0_1024/length-659-level-default-dict-1-8       60.9µs ± 2%     24.9µs ± 0%    -59.15%  (p=0.000 n=9+8)
EncodeAllDict0_1024/length-659-level-better-dict-1-8         182µs ± 2%      188µs ± 3%     +3.24%  (p=0.000 n=10+10)
EncodeAllDict0_1024/length-174-level-fastest-dict-1-8       15.5µs ± 1%      9.4µs ± 2%    -39.17%  (p=0.000 n=10+10)
EncodeAllDict0_1024/length-174-level-default-dict-1-8       56.4µs ± 2%     20.1µs ± 4%    -64.39%  (p=0.000 n=10+10)
EncodeAllDict0_1024/length-174-level-better-dict-1-8         180µs ± 2%      183µs ± 2%     +1.55%  (p=0.011 n=9+9)
```

Data size from 1024 to 8192 bytes:
```
EncodeAllDict1024_8192/length-1076-level-fastest-dict-1-8       19.8µs ± 1%    17.4µs ± 1%   -11.87%  (p=0.000 n=9+10)
EncodeAllDict1024_8192/length-1076-level-default-dict-1-8       81.0µs ± 3%    37.7µs ± 0%   -53.41%  (p=0.000 n=9+8)
EncodeAllDict1024_8192/length-1076-level-better-dict-1-8         223µs ± 2%     219µs ± 2%    -2.13%  (p=0.008 n=9+9)
EncodeAllDict1024_8192/length-5872-level-fastest-dict-1-8       47.0µs ± 2%    47.4µs ± 1%    +0.72%  (p=0.022 n=10+9)
EncodeAllDict1024_8192/length-5872-level-default-dict-1-8        177µs ± 2%     181µs ± 1%    +2.14%  (p=0.000 n=10+10)
EncodeAllDict1024_8192/length-5872-level-better-dict-1-8         359µs ± 0%     356µs ± 3%    -0.95%  (p=0.027 n=8+9)
```

For some reason I can even see some speed up on larger data:
```
EncodeAllDict8192_16384/length-12131-level-fastest-dict-1-8       94.2µs ±28%    77.8µs ± 1%     ~     (p=0.143 n=10+10)
EncodeAllDict8192_16384/length-12131-level-default-dict-1-8        347µs ± 5%     344µs ± 2%     ~     (p=0.762 n=8+10)
EncodeAllDict8192_16384/length-12131-level-better-dict-1-8         635µs ±17%     557µs ± 1%  -12.29%  (p=0.000 n=10+9)
EncodeAllDict8192_16384/length-12131-level-fastest-dict-1#01-8    87.3µs ±27%    78.0µs ± 2%  -10.67%  (p=0.035 n=10+10)
EncodeAllDict8192_16384/length-12131-level-default-dict-1#01-8     388µs ±20%     341µs ± 1%  -11.99%  (p=0.000 n=10+9)
EncodeAllDict8192_16384/length-12131-level-better-dict-1#01-8      650µs ± 9%     549µs ± 1%  -15.57%  (p=0.000 n=8+10)
EncodeAllDict8192_16384/length-9024-level-fastest-dict-1-8        36.9µs ± 8%    36.6µs ± 1%     ~     (p=1.000 n=9+10)
EncodeAllDict8192_16384/length-9024-level-default-dict-1-8         129µs ± 1%     126µs ± 1%   -1.78%  (p=0.000 n=10+10)
EncodeAllDict8192_16384/length-9024-level-better-dict-1-8          316µs ± 6%     299µs ± 1%   -5.39%  (p=0.000 n=9+9)
EncodeAllDict8192_16384/length-12131-level-fastest-dict-1#02-8    78.6µs ± 1%    77.7µs ± 1%   -1.15%  (p=0.015 n=8+9)
EncodeAllDict8192_16384/length-12131-level-default-dict-1#02-8     362µs ± 6%     342µs ± 1%   -5.59%  (p=0.000 n=10+10)
EncodeAllDict8192_16384/length-12131-level-better-dict-1#02-8      611µs ± 6%     553µs ± 2%   -9.38%  (p=0.000 n=10+9)
EncodeAllDict8192_16384/length-12131-level-fastest-dict-1#03-8    83.1µs ±10%    77.6µs ± 0%   -6.64%  (p=0.000 n=10+10)
EncodeAllDict8192_16384/length-12131-level-default-dict-1#03-8     360µs ±10%     343µs ± 1%   -4.84%  (p=0.003 n=9+9)
EncodeAllDict8192_16384/length-12131-level-better-dict-1#03-8      664µs ±13%     548µs ± 1%  -17.34%  (p=0.000 n=10+9)
```

There seem to be some performance gain even with data larger than 16K, but with 64K and more there's no difference at all.
No idea why "better" shows small degradation or improvement in some cases, my modifications shouldn't have affected it any way.