@klauspost 2021-03-23
Interesting. I would never expect randomly branching code to be faster than a straight up memcopy, not even considering the extra writes to update the tables.