@klauspost 2021-03-23
It might as well be a bit more coarsely gained, and bits can be packed, so the actual dirty tables will only be a few hundred bytes.