@klauspost 2021-03-23
@tony2001 I rewrote it slightly for #345 - mainly so regular encodes didn't have to allocate the extra space and waste time writing to the table. Added it to "better" as well. As I also noted in the PR this would also allow to not copy the history when starting a new encode, though that will require some rewriting of the matching code.